Accept mail with charset not set to utf-8.
The plain and HTML views of the mail expected the mail to be encoded as
utf-8. So for mail with other encoding the presentation will be messed
up for characters outside the ascii-range.

This is a working but crude patch since its reencodes the text to utf-8
from whatever encoding it was to start of with. so all messages are
actually displayed as utf-8 messages even if you have a line like this
in your message:
  Content-Type: text/plain; charset="iso-8859-1"